+++ src/components/Button.jsx
@@ -7,6 +7,8 @@
 import React from 'react';
 import styled from 'styled-components';
 
+import { BUTTON_COLORS } from 'constants/colors';
+
 const AnimatedButton = styled.a`
   border-radius: 10px;
   cursor: ${({ disabled }) => (disabled ? 'cursor' : 'pointer')};
@@ -20,46 +22,22 @@ const AnimatedButton = styled.a`
   text-decoration: none;
   transition: all 0.1s;
   user-select: none;
-
   &:active {
     border-bottom: 1px solid;
     transform: translate(0, 5px);
   }
 `;
-
-const colors = {
-  blue: {
-    bgColor: '#428bca',
-    border: '#0a5a9c',
-    color: '#fff',
-  },
-  green: {
-    bgColor: '#5cb85c',
-    border: '#1b7c1b',
-    color: '#fff',
-  },
-  red: {
-    bgColor: '#d9534f',
-    border: '#be2f2b',
-    color: '#fff',
-  },
-  yellow: {
-    bgColor: '#d3ae3f',
-    border: '#8e6c08',
-    color: '#fff',
-  },
-};
 const ColoredAnimatedButton = styled(AnimatedButton)`
-  background-color: ${({ type }) => colors[type].bgColor};
-  border-bottom: 5px solid ${({ type }) => colors[type].border};
-  color: ${({ type }) => colors[type].color};
-  text-shadow: 0px -2px ${({ type }) => colors[type].border};
+  background-color: ${({ type }) => BUTTON_COLORS[type].bgColor};
+  border-bottom: 5px solid ${({ type }) => BUTTON_COLORS[type].border};
+  color: ${({ type }) => BUTTON_COLORS[type].color};
+  text-shadow: 0 -2px ${({ type }) => BUTTON_COLORS[type].border};
 `;
 
 type Props = {|
   disabled?: boolean,
   title: string,
-  type: 'blue' | 'red' | 'green' | 'yellow',
+  type: $Keys<typeof BUTTON_COLORS>,
   handleClick: () => void | (event: SyntheticEvent) => void,
 |};
 export default function ExecuteButton({ handleClick, title, type, disabled }: Props) {