+++ src/constants/colors.js
@@ -5,5 +5,30 @@
  * @flow
  */
 
-export const ACTION = 345;
-export const CONTAINER = 120;
+const WHITE = '#fff';
+
+export const ACTION = '#b63668';
+export const CONTAINER = '#369836';
+
+export const BUTTON_COLORS = {
+  blue: {
+    bgColor: '#428bca',
+    border: '#0a5a9c',
+    color: WHITE,
+  },
+  green: {
+    bgColor: '#5cb85c',
+    border: '#1b7c1b',
+    color: WHITE,
+  },
+  red: {
+    bgColor: '#d9534f',
+    border: '#be2f2b',
+    color: WHITE,
+  },
+  orange: {
+    bgColor: '#D3823F',
+    border: '#B36320',
+    color: WHITE,
+  },
+};